# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ticketing Gamification API
  version: 1.0.0
  description: REST API for ticketing + gamification. All protected routes under `/api/v1` require JWT (and `X-API-Key` if enabled).
servers:
  - url: http://localhost:8080
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional. Required only when API key enforcement is enabled on the server.
  parameters:
    PaginationLimit:
      in: query
      name: limit
      schema:
        type: integer
        default: 50
        maximum: 200
      description: Max items to return (default 50, max 200)
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: validation_error
            message:
              type: string
              example: invalid role
            details:
              description: Optional extra information (e.g., field errors)
              nullable: true
              oneOf:
                - type: object
                - type: array
                  items: {}
                - type: string
    DataResponse:
      type: object
      required: [data]
      properties:
        data: {}
    PagedResponse:
      allOf:
        - $ref: '#/components/schemas/DataResponse'
        - type: object
          properties:
            data:
              type: array
              items: {}
            limit:
              type: integer
...
    TicketPriority:
      type: string
      enum: [low, medium, high, urgent]
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    AuthResponse:
      allOf:
        - $ref: '#/components/schemas/DataResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                token: { type: string }
                refreshToken: { type: string }
                user: { $ref: '#/components/schemas/User' }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    ChangePasswordRequest:
      type: object
      required: [oldPassword, newPassword]
      properties:
        oldPassword: { type: string }
        newPassword: { type: string }
    RegisterRequest:
      type: object
      required: [name, username, password]
      properties:
        name: { type: string }
        username: { type: string }
        password: { type: string }
        role: { type: string }
        avatarUrl: { type: string }
        badges:
          type: array
          items: { type: string }
        bio: { type: string }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        username: { type: string }
        role: { type: string }
        avatarUrl: { type: string }
        badges:
          type: array
          items: { type: string }
        bio: { type: string }
    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        status: { type: string }
        ticketsCount: { type: integer }
        createdAt: { type: string, format: date-time }
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            members:
              type: array
              items:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  name: { type: string }
                  role: { type: string }
            invites:
              type: array
              items:
                $ref: '#/components/schemas/Invite'
            activity:
              type: array
              items:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  text: { type: string }
                  timestamp: { type: string, format: date-time }
...
    GamificationStats:
      type: object
      properties:
        userId: { type: string, format: uuid }
        xpTotal: { type: integer }
        level: { type: integer }
        nextLevelThreshold: { type: integer }
        ticketsClosed: { type: integer }
        streakDays: { type: integer }
        lastTicketClosedAt: { type: string, format: date-time }
    GamificationEvent:
      type: object
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        ticketId: { type: string, format: uuid, nullable: true }
        priority: { $ref: '#/components/schemas/TicketPriority' }
        xp: { type: integer }
        note: { type: string }
        createdAt: { type: string, format: date-time }
...
    Invite:
      type: object
      properties:
        code: { type: string }
        maxUses: { type: integer }
        uses: { type: integer }
        expiresAt: { type: string, format: date-time }
    LeaderboardEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        username: { type: string }
        role: { type: string }
        xpTotal: { type: integer }
        level: { type: integer }
        ticketsClosedCount: { type: integer }
        rank: { type: integer }
        xpGap: { type: integer }
paths:
  /api/v1/auth/login:
    post:
      summary: Login with username/password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/auth/register:
    post:
      summary: Register new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/auth/refresh:
    post:
      summary: Refresh access token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: Refreshed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/auth/change-password:
    post:
      summary: Change password (authenticated)
      tags: [Auth]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '204': { description: Updated }
        '401':
          description: Invalid old password
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /api/v1/projects:
    get:
      summary: List projects
      tags: [Projects]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PaginationLimit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Project' }
                      limit: { type: integer }
    post:
      summary: Create project
      tags: [Projects]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, description]
              properties:
                name: { type: string }
                description: { type: string }
                members:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectDetail' }
  /api/v1/projects/{id}:
    get:
      summary: Project detail
      tags: [Projects]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectDetail' }
  /api/v1/projects/{id}/members:
    post:
      summary: Add project member
      tags: [Projects]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId: { type: string, format: uuid }
                role: { type: string }
      responses:
        '204': { description: Added }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/projects/{id}/invites:
    post:
      summary: Generate invite code
      tags: [Projects]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [maxUses, expiryDays]
              properties:
                maxUses: { type: integer }
                expiryDays: { type: integer }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Invite' }
  /api/v1/projects/join:
    post:
      summary: Join project via invite code
      tags: [Projects]
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
      responses:
        '200':
          description: Joined
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectDetail' }
...
  /api/v1/gamification/stats/{userId}:
    get:
      summary: Gamification stats
      tags: [Gamification]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GamificationStats' }
        '404':
          description: Stats not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/gamification/events:
    get:
      summary: XP events feed
      tags: [Gamification]
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: userId
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/PaginationLimit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/GamificationEvent' }
                      limit: { type: integer }
  /api/v1/gamification/leaderboard:
    get:
      summary: Leaderboard
      tags: [Gamification]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PaginationLimit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/LeaderboardEntry' }
...